Save chunk groups, make PresenceData concurrent-able
This comes with the penalty of me using AtomicIntegers,
but I think in the end it might be better than the thing we had
before as this method drastically cuts the amounts of stuff
that is allocated (especially for auto-(un)-boxing).

I still haven't tested this all, but I am confident that it works
to some degree.