Implement explosion protection and prevent tamed entities from killing other entities
I'd need to test the performance impact of this but I think it should not be all too
taxing. And this should also be one of the last items on my list of stuff to implement.

I do not think that it makes sense to prohibit lava & water movement by default and
I do not think that people are going to enable this feature, so it is going to be wasted
processing power. So the water/lava flow protection will probably not be implemented as of now.

On the topic of dynmap/bluemap integration - I have planned it but just not yet as I do not
make use of that plugin in my production server.