Serialisation unit tests
This uncovered a bug with successor serialisation.
However I am sure that it worked beforehand, so I am not sure why it is bugged in the first place.

Furthermore there seems to be a big performance sink somewhere. I guess I have to profile
it and check whether this performance loss is intrinsic to the RegionatedIntIntToObjectMap to whether it is
caused by serialisation. Perhaps it is a mix of both. We will see